numerous improvements
enabled full 16bpc when hdr is supported and on
*numerous* graphics bugs fixed in volumetrics and reflections
toying with dynamic camera height, may remove....
 texturing improvements supporting ktx2 & ktx
support independent width and height for whole world volume dimensions
numerous improvements

+++ MinCity/lightBuffer3D.h
@@ -345,15 +345,15 @@ public:
 	{
 		writeonlyBuffer3D<XMFLOAT4A, LightWidth, LightHeight, LightDepth>::clear_tracking();
 
-		uint32_t size(0);
+		size_t size(0);
 
 		size = LightWidth * LightHeight * LightDepth * sizeof(std::atomic_uint32_t);
 		_internal = (std::atomic_uint32_t * __restrict)scalable_aligned_malloc(size, CACHE_LINE_BYTES); // *bugfix - over-aligned for best performance and to prevent false sharing
-		_block_size_atomic = size_t(size) / hardware_concurrency;
+		_block_size_atomic = (uint32_t)(size / hardware_concurrency);
 
 		size = LightWidth * LightHeight * LightDepth * sizeof(XMFLOAT4A);
 		_cache = (XMFLOAT4A * __restrict)scalable_aligned_malloc(size, CACHE_LINE_BYTES); // *bugfix - over-aligned for best performance (wc copy) and to prevent false sharing
-		_block_size_cache = size_t(size) / hardware_concurrency;
+		_block_size_cache = (uint32_t)(size / hardware_concurrency);
 
 		clear(0);
 	}
